+++ cypress/fixtures/empty.json
@@ -1,0 +1,8 @@
+{
+  "customers": [],
+  "pageInfo": {
+    "currentPage": 1,
+    "totalPages": 1,
+    "totalCustomers": 0
+  }
+}
