+++ cypress/fixtures/tecnology/smallTechnology.json
