+++ cypress/fixtures/withAddress.json
@@ -1,0 +1,27 @@
+{
+  "customers": [
+    {
+      "id": 1,
+      "name": "Mathews Co",
+      "employees": 99,
+      "industry": "Logistics",
+      "contactInfo": {
+        "name": "John Smith",
+        "email": "jsmith@americasinc.com"
+      },
+      "address": {
+          "street": "988 Kimberly Fort Apt. 921",
+          "city": "Lake Tracy",
+          "state": "Connecticut",
+          "zipCode": "07115",
+          "country": "United States of America"
+      },
+      "size": "Small"
+    }
+  ],
+  "pageInfo": {
+    "currentPage": 1,
+    "totalPages": 1,
+    "totalCustomers": 1
+  }
+}
