+++ cypress/support/commands.js
@@ -1,0 +1,3 @@
+Cypress.Commands.add('getByClassThatStartsWith', classPart => {
+  cy.get(`[class^="${classPart}"]`);
+});
