+++ cypress/support/e2e.js
@@ -1,4 +1,3 @@
-import './api-commands'
-import './gui-commands'
-
 import 'cypress-axe';
+
+import './commands'
