--- cypress/support/gui-commands.js
@@ -1,3 +1,0 @@
-Cypress.Commands.add('getByClassThatStartsWith', (classPart) => {
-  cy.get(`[class^="${classPart}"]`);
-});
